:truck: Move data_removal into separate app
