+++ src/openforms/data_removal/__init__.py
@@ -1,0 +1,6 @@
+"""
+Provide functionality to remove (submission) data.
+
+GDPR regulations require data to be removed or be made un-identifiable. This sub-package
+provides the plumbing for that.
+"""
