+++ src/openforms/submissions/constants.py
@@ -13,12 +13,3 @@ class RegistrationStatuses(DjangoChoices):
     in_progress = ChoiceItem("in_progress", _("In Progress (not registered yet)"))
     success = ChoiceItem("success", _("Success"))
     failed = ChoiceItem("failed", _("Failed"))
-
-
-class RemovalMethods(DjangoChoices):
-    delete_permanently = ChoiceItem(
-        "delete_permanently", _("Submissions will be deleted")
-    )
-    make_anonymous = ChoiceItem(
-        "make_anonymous", _("Sensitive data within the submissions will be deleted")
-    )
